requirements.txt
# requirements.txt
spotipy
google-api-python-client
python-dotenv